modified register validation and hide some links on responsive

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,7 +20,7 @@ class CreateUsersTable extends Migration
             $table->string('avatar_id')->default('');
             $table->string('firstname');
             $table->string('lastname');
-            $table->integer('siret');
+            $table->bigInteger('siret');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
